Refactor:setup for auth context
