+++ src/Context/DataContext/index.js
@@ -1,0 +1,4 @@
+export { useData } from "./useData";
+export { DataProvider } from "./DataProvider";
+export * as Actions from "./ActionTypes";
+export { products } from "./Products";
