+++ src/Context/index.js
@@ -1,4 +1,1 @@
-export { useData } from "./useData";
-export { DataProvider } from "./DataProvider";
-export * as Actions from "./ActionTypes";
-export { products } from "./Products";
+export { useData, DataProvider, Actions } from "./DataContext";
